✨ (Various): Implement new features and enhancements
🐛 (InfoController): Add authorization for super admin faculty majors
✨ (InfoController): Add endpoint to get super admin faculty majors
✨ (ScheduleController): Add major year parameter to get admin schedule by month
✨ (ScheduleManagementController): Add schedule id to day schedule dto
✨ (SubjectController): Update SaveFileAsync method call in SubjectController
✨ (AuthMappings): Add number of years to admin dto
✨ (BaseAdminDto): Add NumberOfYears property to BaseAdminDto
✨ (InfoMappings): Add number of years to major name dto
✨ (MajorNameDto): Add NumberOfYears property to MajorNameDto
✨ (DayScheduleDto): Add ScheduleId property to DayScheduleDto
✨ (ScheduleMappings): Add ScheduleId to DayScheduleDto
✨ (UnifiedSubjectDto): Change MidtermGrade and FinalGrade to double
🐛 (DatabaseSeeder): Comment out ClearApplicationDataAsync method call
✨ (Subject): Change MidtermGrade and FinalGrade to double
⬆ (Dependencies): Update database migration
✨ (Program): Enable static file serving from wwwroot

The InfoController was updated to include an endpoint for retrieving majors associated with a super admin's faculty.
The ScheduleController now requires a major year parameter for retrieving the admin schedule by month.
The ScheduleManagementController now includes the schedule ID in the DayScheduleDto.
The SubjectController's SaveFileAsync method call was updated.
The AdminDto and MajorNameDto now include the number of years for the major.
The DayScheduleDto now includes the ScheduleId.
The UnifiedSubjectDto now uses double for MidtermGrade and FinalGrade.
The database migration was updated to reflect the changes in grade types.
Static file serving from wwwroot is now enabled.

+++ UniSphere/DTOs/Subjects/UnifiedSubjectDto.cs
@@ -9,8 +9,8 @@ public sealed record UnifiedSubjectDto
     public Guid? LabId { get; init; }
     public required int Year { get; init; }
     public required int Semester { get; init; }
-    public required int MidtermGrade { get; init; }
-    public required int FinalGrade { get; init; }
+    public required double MidtermGrade { get; init; }
+    public required double FinalGrade { get; init; }
     public required bool IsLabRequired { get; init; }
     public required bool IsMultipleChoice { get; init; }
     public required bool IsOpenBook { get; init; }
